+++ httpcore/_sync/base.py
@@ -2,7 +2,7 @@ import enum
 from types import TracebackType
 from typing import Iterator, List, Tuple, Type
 
-from .._types import URL, Headers, TimeoutDict
+from .._types import URL, Headers, T, TimeoutDict
 
 
 class NewConnectionRequired(Exception):
@@ -75,8 +75,8 @@ class SyncHTTPTransport:
         **Parameters:**
 
         * **method** - `bytes` - The HTTP method, such as `b'GET'`.
-        * **url** - `Tuple[bytes, bytes, Optional[int], bytes]` - The URL as a 4-tuple of
-        (scheme, host, port, path).
+        * **url** - `Tuple[bytes, bytes, Optional[int], bytes]` - The URL as a 4-tuple
+        of (scheme, host, port, path).
         * **headers** - `Optional[List[Tuple[bytes, bytes]]]` - Any HTTP headers
         to send with the request.
         * **stream** - `Optional[SyncByteStream]` - The body of the HTTP request.
@@ -106,7 +106,7 @@ class SyncHTTPTransport:
         and any keep alive connections.
         """
 
-    def __enter__(self) -> "SyncHTTPTransport":
+    def __enter__(self: T) -> T:
         return self
 
     def __exit__(