+++ tests/async_tests/test_interfaces.py
@@ -5,6 +5,13 @@ import pytest
 
 import httpcore
 from httpcore._types import URL
+from tests.conftest import Server
+from tests.utils import lookup_async_backend
+
+
+@pytest.fixture(params=["auto", "anyio"])
+def backend(request):
+    return request.param
 
 
 async def read_body(stream: httpcore.AsyncByteStream) -> bytes:
@@ -17,16 +24,16 @@ async def read_body(stream: httpcore.AsyncByteStream) -> bytes:
         await stream.aclose()
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_http_request() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_http_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -34,16 +41,16 @@ async def test_http_request() -> None:
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_https_request() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_https_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -51,9 +58,9 @@ async def test_https_request() -> None:
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_request_unsupported_protocol() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_request_unsupported_protocol(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"ftp", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
@@ -61,16 +68,16 @@ async def test_request_unsupported_protocol() -> None:
             await http.arequest(method, url, headers)
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_http2_request() -> None:
-    async with httpcore.AsyncConnectionPool(http2=True) as http:
+@pytest.mark.anyio
+async def test_http2_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend, http2=True) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/2"
         assert status_code == 200
@@ -78,16 +85,16 @@ async def test_http2_request() -> None:
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_closing_http_request() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_closing_http_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org"), (b"connection", b"close")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -95,16 +102,16 @@ async def test_closing_http_request() -> None:
         assert url[:3] not in http._connections  # type: ignore
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_http_request_reuse_connection() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_http_request_reuse_connection(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -117,7 +124,7 @@ async def test_http_request_reuse_connection() -> None:
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -125,16 +132,16 @@ async def test_http_request_reuse_connection() -> None:
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_https_request_reuse_connection() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_https_request_reuse_connection(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -147,7 +154,7 @@ async def test_https_request_reuse_connection() -> None:
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -155,16 +162,16 @@ async def test_https_request_reuse_connection() -> None:
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
-@pytest.mark.usefixtures("async_environment")
-async def test_http_request_cannot_reuse_dropped_connection() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+@pytest.mark.anyio
+async def test_http_request_cannot_reuse_dropped_connection(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -173,7 +180,7 @@ async def test_http_request_cannot_reuse_dropped_connection() -> None:
 
         # Mock the connection as having been dropped.
         connection = list(http._connections[url[:3]])[0]  # type: ignore
-        connection.is_connection_dropped = lambda: True
+        connection.is_connection_dropped = lambda: True  # type: ignore
 
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
@@ -181,7 +188,7 @@ async def test_http_request_cannot_reuse_dropped_connection() -> None:
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -190,36 +197,43 @@ async def test_http_request_cannot_reuse_dropped_connection() -> None:
 
 
 @pytest.mark.parametrize("proxy_mode", ["DEFAULT", "FORWARD_ONLY", "TUNNEL_ONLY"])
-@pytest.mark.usefixtures("async_environment")
-async def test_http_proxy(proxy_server: URL, proxy_mode: str) -> None:
+@pytest.mark.anyio
+async def test_http_proxy(proxy_server: URL, proxy_mode: str, backend: str) -> None:
     method = b"GET"
     url = (b"http", b"example.org", 80, b"/")
     headers = [(b"host", b"example.org")]
     max_connections = 1
     async with httpcore.AsyncHTTPProxy(
-        proxy_server, proxy_mode=proxy_mode, max_connections=max_connections,
+        proxy_server,
+        proxy_mode=proxy_mode,
+        max_connections=max_connections,
+        backend=backend,
     ) as http:
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
         assert reason == b"OK"
 
 
-@pytest.mark.asyncio
-# This doesn't run with trio, since trio doesn't support local_address.
-async def test_http_request_local_address() -> None:
-    async with httpcore.AsyncConnectionPool(local_address="0.0.0.0") as http:
+@pytest.mark.anyio
+async def test_http_request_local_address(backend: str) -> None:
+    if backend == "auto" and lookup_async_backend() == "trio":
+        pytest.skip("The trio backend does not support local_address")
+
+    async with httpcore.AsyncConnectionPool(
+        backend=backend, local_address="0.0.0.0"
+    ) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
         http_version, status_code, reason, headers, stream = await http.arequest(
             method, url, headers
         )
-        body = await read_body(stream)
+        await read_body(stream)
 
         assert http_version == b"HTTP/1.1"
         assert status_code == 200
@@ -229,10 +243,10 @@ async def test_http_request_local_address() -> None:
 
 # mitmproxy does not support forwarding HTTPS requests
 @pytest.mark.parametrize("proxy_mode", ["DEFAULT", "TUNNEL_ONLY"])
-@pytest.mark.usefixtures("async_environment")
 @pytest.mark.parametrize("http2", [False, True])
+@pytest.mark.anyio
 async def test_proxy_https_requests(
-    proxy_server: URL, ca_ssl_context: ssl.SSLContext, proxy_mode: str, http2: bool,
+    proxy_server: URL, ca_ssl_context: ssl.SSLContext, proxy_mode: str, http2: bool
 ) -> None:
     method = b"GET"
     url = (b"https", b"example.org", 443, b"/")
@@ -279,12 +293,16 @@ async def test_proxy_https_requests(
         (True, 0.0, {"https://example.org": ["HTTP/2, ACTIVE, 2 streams"]}, {},),
     ],
 )
-@pytest.mark.usefixtures("async_environment")
+@pytest.mark.anyio
 async def test_connection_pool_get_connection_info(
-    http2, keepalive_expiry, expected_during_active, expected_during_idle
+    http2: bool,
+    keepalive_expiry: float,
+    expected_during_active: dict,
+    expected_during_idle: dict,
+    backend: str,
 ) -> None:
     async with httpcore.AsyncConnectionPool(
-        http2=http2, keepalive_expiry=keepalive_expiry
+        http2=http2, keepalive_expiry=keepalive_expiry, backend=backend
     ) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
@@ -311,11 +329,13 @@ async def test_connection_pool_get_connection_info(
     platform.system() not in ("Linux", "Darwin"),
     reason="Unix Domain Sockets only exist on Unix",
 )
-@pytest.mark.usefixtures("async_environment")
-async def test_http_request_unix_domain_socket(uds_server) -> None:
+@pytest.mark.anyio
+async def test_http_request_unix_domain_socket(
+    uds_server: Server, backend: str
+) -> None:
     uds = uds_server.config.uds
     assert uds is not None
-    async with httpcore.AsyncConnectionPool(uds=uds) as http:
+    async with httpcore.AsyncConnectionPool(uds=uds, backend=backend) as http:
         method = b"GET"
         url = (b"http", b"localhost", None, b"/")
         headers = [(b"host", b"localhost")]
@@ -327,3 +347,48 @@ async def test_http_request_unix_domain_socket(uds_server) -> None:
         assert reason == b"OK"
         body = await read_body(stream)
         assert body == b"Hello, world!"
+
+
+@pytest.mark.parametrize("max_keepalive", [1, 3, 5])
+@pytest.mark.parametrize("connections_number", [4])
+@pytest.mark.anyio
+async def test_max_keepalive_connections_handled_correctly(
+    max_keepalive: int, connections_number: int, backend: str
+) -> None:
+    async with httpcore.AsyncConnectionPool(
+        max_keepalive_connections=max_keepalive, keepalive_expiry=60, backend=backend
+    ) as http:
+        method = b"GET"
+        url = (b"http", b"example.org", 80, b"/")
+        headers = [(b"host", b"example.org")]
+
+        connections_streams = []
+        for _ in range(connections_number):
+            _, _, _, _, stream = await http.request(method, url, headers)
+            connections_streams.append(stream)
+
+        try:
+            for i in range(len(connections_streams)):
+                await read_body(connections_streams[i])
+        finally:
+            stats = await http.get_connection_info()
+
+            connections_in_pool = next(iter(stats.values()))
+            assert len(connections_in_pool) == min(connections_number, max_keepalive)
+
+
+@pytest.mark.anyio
+async def test_explicit_backend_name() -> None:
+    async with httpcore.AsyncConnectionPool(backend=lookup_async_backend()) as http:
+        method = b"GET"
+        url = (b"http", b"example.org", 80, b"/")
+        headers = [(b"host", b"example.org")]
+        http_version, status_code, reason, headers, stream = await http.request(
+            method, url, headers
+        )
+        await read_body(stream)
+
+        assert http_version == b"HTTP/1.1"
+        assert status_code == 200
+        assert reason == b"OK"
+        assert len(http._connections[url[:3]]) == 1  # type: ignore